make interface for response builder
