+++ src/main/java/action/DoubleAction.java
@@ -4,15 +4,9 @@ import model.Card;
 import request.DoubleRequest;
 import request.StandRequest;
 import response.BaseResponse;
-import responseBuilder.ResponseBuilder;
-import util.BlackjackUtils;
 import util.GameState;
 
-import java.util.ArrayList;
 import java.util.List;
-
-import static util.CommonConstants.DEALER_STANDING_POINTS;
-import static util.CommonConstants.SECOND;
 
 public class DoubleAction extends BaseAction <BaseResponse, DoubleRequest> {
     @Override
